♻️ Refactor: strongly-type UserErrors
