+++ packages/backend/src/sync/controllers/sync.controller.ts
@@ -2,8 +2,8 @@ import { Request } from "express";
 import { Status } from "@core/errors/status.codes";
 import { Logger } from "@core/logger/winston.logger";
 import { Payload_Sync_Notif } from "@core/types/sync.types";
-import { UserError } from "@backend/common/constants/error.constants";
 import { SyncError } from "@backend/common/errors/sync/sync.errors";
+import { UserError } from "@backend/common/errors/user/user.errors";
 import {
   isFullSyncRequired,
   isInvalidGoogleToken,
